+++ node/constants/index.ts
@@ -1,5 +1,6 @@
 export const APP_NAME = 'vtex.cross-device-cart'
 export const BUCKET = 'cross_device_cart'
+export const DEFAULT_ORDER_FORM_ID = 'default-order-form'
 export const SETTINGS_PATH = 'app_settings'
 export const DEFAULT_SETTINGS = {
   isAutomatic: true,